Home, sheet, room pages and their operations are all inaccessible for a unauthenticated user. Users can only see their own sheets unless they are a superuser, then they can see all sheets. Additional formatting in HTML and javaScript was added to calculate changes of modifiers for character sheets and for better organization. Refactoring of the CRUD operation method names in view.py. Added new documentation throughout the project. Sheet now automatically adds the user to its field, users no longer can assign a user to a sheet (unless in the admin panel).
Spike Epic #1 should now be complete with the addition of storing features and spells in a textfield. As a user can now create a sheet, add statistics, options and notes to the character and save it to their account. As stated by US1. The user can also now modify statistics of the character, remove or insert new character options/notes and remove the character sheet from their account as stated by US2.

Fixed an issue where page would 404 after sheet creation even though sheet was properly created as expected.

+++ base/views.py
@@ -1,11 +1,14 @@
 from multiprocessing import context
+
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 from django.contrib import messages
-from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login as dj_login, logout
 from django.contrib.auth.forms import UserCreationForm
+
 from .models import *
 from .forms import SheetForm
+
 
 # Create your views here.
 
@@ -16,16 +19,16 @@ def regPage(request):
     if request.method == 'POST':
         form = UserCreationForm(request.POST)
         if form.is_valid():
-            user = form.save(commit = False)
+            user = form.save(commit=False)
             user.username = user.username.lower()
             user.save()
             return redirect('home')
         else:
-            messages.error(request, 'An error occured during')
-
+            messages.error(request, 'An error occurred during')
 
     context = {'form': form}
     return render(request, 'base/user_registration.html', context)
+
 
 # the view used by the login page
 def login(request):
@@ -36,14 +39,14 @@ def login(request):
 
         # check if the user does exits
         try:
-            user = User.objects.get(username = username)
+            user = User.objects.get(username=username)
         except:
             messages.error(request, 'User does not exist')
 
         # validate the user login information
-        user = authenticate(request, username = username, password = password)
+        user = authenticate(request, username=username, password=password)
 
-        #if the user is not None, then login the user and redirect
+        # if the user is not None, then login the user and redirect
         # them to the home page
         if user is not None:
             dj_login(request, user)
@@ -54,10 +57,13 @@ def login(request):
     context = {}
     return render(request, 'base/login.html', context)
 
+
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
+
+@login_required(login_url='/login')
 def home(request):
     rooms = Room.objects.all()
     sheets = Sheet.objects.all()
@@ -65,23 +71,38 @@ def home(request):
     return render(request, 'base/home.html', context)
 
 
+@login_required(login_url='/login')
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='/login')
 def sheet(request, pk):
     sheet = Sheet.objects.get(id=pk)
+
+    if request.user != sheet.user and not request.user.is_superuser:  # only admins or users can view their own sheets
+        return redirect('home')
+
     context = {'sheet': sheet}
     return render(request, 'base/sheet.html', context)
 
 
-def createSheet(request):
-    form = SheetForm()
+@login_required(login_url='/login')  # can only be used by users that are logged in, else redirect to login page
+def create_sheet(request):
+    """
+    Create sheet and save if valid.
+    :param request: object to pass state to the function
+    :return: url to redirect the user to
+    """
+    form = SheetForm()  # get default form
+
     if request.method == 'POST':
-        form = SheetForm(request.POST)
-        if form.is_valid():
+        form = SheetForm(request.POST)  # get form for post
+        if form.is_valid():  # if valid, save, then redirect
+            form = form.save(commit=False)  # save the form, then update user to user who created the sheet
+            form.user = request.user
             form.save()
             return redirect('home')
 
@@ -89,22 +110,42 @@ def createSheet(request):
     return render(request, 'base/sheet_form.html', context)
 
 
-def updateSheet(request, pk):
-    sheet = Sheet.objects.get(id=pk)
-    form = SheetForm(initial=sheet)
+@login_required(login_url='/login')  # can only be used by users that are logged in, else redirect to login page
+def update_sheet(request, pk):
+    """
+    Edit the sheet and saves the data if valid.
+    :param request: object to pass state to the function
+    :param pk: ID for the sheet
+    :return: url to redirect the user to
+    """
+    sheet = Sheet.objects.get(id=pk)  # get sheet with ID
+    form = SheetForm(instance=sheet)  # get current form
+
+    if request.user != sheet.user and not request.user.is_superuser:  # only admins or users can edit their own sheets
+        return redirect('home')
 
     if request.method == 'POST':
-        form = SheetForm(request.POST, instance=sheet)
-        if form.is_valid():
+        form = SheetForm(request.POST, instance=sheet)  # get form for post
+        if form.is_valid():  # if valid, save, then redirect to previous location
             form.save()
-            return redirect('home')
+            return redirect(f'/sheet/{pk}')
 
-    context = {}
-    return render(request, 'base/sheet_form.html', context)
+    context = {'form': form}
+    return render(request, 'base/sheet_form.html', context)  # render existing form
 
 
-def deleteSheet(request, pk):
+@login_required(login_url='/login')  # can only be used by users that are logged in, else redirect to login page
+def delete_sheet(request, pk):
+    """
+    Delete sheet on user confirmation.
+    :param request: object to pass state to the function
+    :return: url to redirect the user to
+    """
     sheet = Sheet.objects.get(id=pk)
+
+    if request.user != sheet.user and not request.user.is_superuser:  # only admins or users can delete their own sheets
+        return redirect('home')
+
     if request.method == 'POST':
         sheet.delete()
         return redirect('home')
